requirements.txt
pytest
sphinx
sphinx-rtd-theme
pyinstaller
auto-py-to-exe
pygame
pillow